requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
Flask-Cors==3.0.10
qrcode[pil]==7.3.1
pymongo==4.10.1
reportlab==4.2.5